requirements.txt
duckdb
pandas
requests
python-dotenv
jupyter
ipykernel
papermill
nbconvert
xlsxwriter
pillow